Refactor transaction handling in inpage providers to improve clarity and consistency
- Updated import paths for RequestInput in multiple providers to align with the new structure, enhancing maintainability.
- Introduced a new GasFeeAdjuster component for managing gas fee settings in transactions, improving user experience during transaction signing.
- Added new interfaces and utility functions for Solana transaction parsing and simulation, streamlining the transaction processing workflow.
- Enhanced the SendTxOverview component to better handle transaction details and user interactions, ensuring a smoother experience.
