+++ clients/extension/src/inpage/providers/cosmos.ts
@@ -1,6 +1,6 @@
 import { getCosmosChainByChainId } from '@core/chain/chains/cosmos/chainInfo'
 import { callBackground } from '@core/inpage-provider/background'
-import { RequestInput } from '@core/inpage-provider/tx/temp/interfaces'
+import { RequestInput } from '@core/inpage-provider/popup/view/resolvers/sendTx/interfaces'
 import { NotImplementedError } from '@lib/utils/error/NotImplementedError'
 
 import { EIP1193Error } from '../../background/handlers/errorHandler'
