+++ clients/extension/src/inpage/providers/ethereum.ts
@@ -8,7 +8,7 @@ import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
 import { callBackground } from '@core/inpage-provider/background'
 import { callPopup } from '@core/inpage-provider/popup'
 import { Eip712V4Payload } from '@core/inpage-provider/popup/interface'
-import { RequestInput } from '@core/inpage-provider/tx/temp/interfaces'
+import { RequestInput } from '@core/inpage-provider/popup/view/resolvers/sendTx/interfaces'
 import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
 import { attempt, withFallback } from '@lib/utils/attempt'
 import { NotImplementedError } from '@lib/utils/error/NotImplementedError'
