+++ clients/extension/src/inpage/providers/polkadot.ts
@@ -1,5 +1,5 @@
 import { OtherChain } from '@core/chain/Chain'
-import { RequestInput } from '@core/inpage-provider/tx/temp/interfaces'
+import { RequestInput } from '@core/inpage-provider/popup/view/resolvers/sendTx/interfaces'
 import { NotImplementedError } from '@lib/utils/error/NotImplementedError'
 import EventEmitter from 'events'
