+++ core/inpage-provider/popup/view/resolvers/sendTx/core/solana/simulate/index.ts
