+++ core/inpage-provider/popup/view/resolvers/sendTx/interfaces.ts
